Add Hangman.Game.make_move() function
This function needs to check whether the guess has already been used or not.
This cannot be done using pattern matching or a guard clause. So we will delegate
the work to another function containing one extra argument, the fact that the
guess is already in the used MapSet or not.
Doing this avoids to write a conditional in our make_move() function.